#64 Add basic timeline structure with special dependency system
Add the basic structure of the TimelineWidget. It uses a custom
dependency system for scrolling positions and widget sizes.
Currently the widget has two parts, the toolbar part and the scrollable
timeline part which shows the tracks. The scrollable part is made up of
three scroll areas and two separate scroll bars. These are all linked
together in different ways so that if one view is scrolled all the
others adjust automatically. This guarantees that parts which belong
together aren't shifted against each other.
Also, there are dependencies implemented which control the sizes of
different parts and scroll areas. The reason for that is simply to
prevent scroll areas from covering their widget in one direction and
activating scrolling to compensate that. In particular: One of the three
scroll areas shouldn't scroll horizontally and one of them shouldn't
scroll vertically.
Because there are currently no contents available for the timeline the
classes come with a little bit of extra code making the behaviour of the
timeline visible.
Due to the short time remaining until some results have to be visible to
other developers the code isn't perfectly structured. Basically,
everything works efficiently and fine. But there is a bit of "spaghetti
code". Also, the Widget isn't added to the main GUI yet and comes with
no direct way to run it. This will be fixed very soon.