#300 Remove redundant code of scrollbar connecting methods and cleanup
Merge connect and disconnect methods for different orientations of
scrolling to remove redundant code. Accordingly implement the choice
completely new by adding an orientation parameter to the connect and
disconnect methods.
Remove the automatic shutoff of the ConnectableScrollArea's scroll bars
for the sake of better predictability.
Adapt the usages in the project to the new methods.