#263 Fix slider lengths in TimelineScrollArea
Add synchronisation between the viewport size of the
ConnectableScrollArea and its external scroll bar's page steps. Because
the missing synchronisation was the reason for the bug it could be fixed
this way.
Achieved by emitting a Qt signal with the viewport's width or height if
the scroll area's size changes and the connection of this signal to the
scroll bars which get connected to the scroll area.