#129 First changes for timeline zoom (contains bugs)
An unstructured set of changes to the timeline which consists of crucial
corrections of the model view pattern and of changes to make timeables
zoomable.
The model view pattern of the timeline is not implemented well. It's not
a good idea to build on top of these flaws. Also because the timeline
controller doesn't really separate view and controller a change to one
side leads to the need of changing the other side, too. So implementing
timeline zoom requires vast refactoring in all the timeline model, the
timeline controller and the timeline model.
Therefore it's a huge task to do all the related changes. Because of
that this commit only contains a small fraction of all necessary
modifications. This leads to a few bugs which will be hopefully fixed by
making the rest of the changes.