#360 Intermediate state of timeline refactoring (not executable)
Mostly extensive changes to the model part of the timeline. Implement
a new class TrackModel with subclasses VideoTrack and AudioTrack to
have easier and more flexible management of tracks than Openshot
provides. Build up a structure where a TimelineModel consists of
multiple TrackModels which themselves consist of multiple
TimeableModels.
All the TimelineModel, the TrackModels and the TimeableModels know the
TimelineController which the timeline belongs to. This makes it easier
to communicate changes to the controller.

It is not guaranteed that everything is already implemented on the model
side. This is due to the problem that the original implementation had
many design pattern flaws and it's not very easy to get a complete
overview over all the things to be changed. This version of UbiCast
won't work properly.