#360 Intermediate state of timeline refactoring, buggy, not complete
Massive changes and reimplementations of the whole model of the
timeline and of the TimelineController, the TrackView and the
TimeableView.
The timeline data has now been completely (as far as can be seen) moved
from the view to the model. Operations don't change the objects in the
view anymore but only work on the model. The model then tells its
changes to the TimelineController which itself says the TimelineView
what to display.
The program is now runnable again and can create and save projects.
The first operations work already but may have bugs.

Still, there are many communication routines to implement. Also, drag
and drop in the timeline needs some quite extensive work to make it more
reliable and give proper feedback to the user.